Initial push to the repo
Create the structure of the project including directories and
class files. Every class files in this project has been initialized
to include the shell of the methods that need to be written and
in case of the `Histogram` class, include the template structure
that was given by the professor.

PS: The `.idea` directory is added tentatively until further
determination of the worthiness of its inclusion/exlusion.

Changes to be committed:
	new file:   .gitignore
	new file:   .idea/misc.xml
	new file:   .idea/modules.xml
	new file:   .idea/vcs.xml
	modified:   README.md
	new file:   monte-carlo.iml
	new file:   src/Histogram.java
	new file:   src/Metrics.java
	new file:   src/MonteCarlo.java
	new file:   src/Simulation.java